+++ sample/src/com/manuelpeinado/multichoiceadapter/demo/HomeActivity.java
@@ -1,0 +1,43 @@
+package com.manuelpeinado.multichoiceadapter.demo;
+
+import java.util.Arrays;
+import java.util.List;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.ArrayAdapter;
+import android.widget.ListView;
+
+import com.actionbarsherlock.app.SherlockListActivity;
+
+public class HomeActivity extends SherlockListActivity {
+    private List<ActivityInfo> activitiesInfo = Arrays.asList(
+            new ActivityInfo(BasicUsageActivity.class, R.string.activity_title_basic_usage),
+            new ActivityInfo(CheckboxItemsActivity.class, R.string.activity_title_checkbox_items));
+    
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+        String[] titles = getActivityTitles();
+        setListAdapter(new ArrayAdapter<String>(
+                this, android.R.layout.simple_list_item_1, android.R.id.text1, titles));
+    }
+    
+    @Override
+    protected void onListItemClick(ListView l, View v, int position, long id) {
+        Class<? extends BaseActivity> class_ = activitiesInfo.get(position).activityClass;
+        Intent intent = new Intent(this, class_);
+        startActivity(intent);
+    }
+
+    private String[] getActivityTitles() {
+        String[] result = new String[activitiesInfo.size()];
+        int i = 0;
+        for (ActivityInfo info : activitiesInfo) {
+            result[i++] = getString(info.titleResourceId);
+        }
+        return result;
+    }
+}