+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version 1.0.2 *(2013-03-24)*
  * Fixed a bug which prevented the selected item background from being shown the second time items were selected
  * Simplified the code that deals with item removal in the sample app
  * Added support for checkboxes
+ * Added support for multiple activities to the sample app
 
 Version 1.0.1 *(2013-03-23)*
 ----------------------------